1. 预下载、下载、缓冲大小、预加载配置的类型修改为浮点类型 2. 点播预加载、缓冲大小增加默认值 3. 规避在flutter高版本中，ios app在后台停留时间过长导致app被杀的时候，如果app还一直在离线下载，会造成app资源释放不一致问题，导致flutter播放器插件engine已经结束，但是下载进程还在回调，产生了crash (merge request !256)
Squash merge branch 'kongdywang/main/fix_ios_download_crash_when_kill' into 'main'
1. 预下载、下载、缓冲大小、预加载配置的类型修改为浮点类型
2. 点播预加载、缓冲大小增加默认值
3. 规避在flutter高版本中，ios app在后台停留时间过长导致app被杀的时候，如果app还一直在离线下载，会造成app资源释放不一致问题，导致flutter播放器插件engine已经结束，但是下载进程还在回调，产生了crash

+++ Flutter/lib/Core/txplayer_messages.dart
@@ -351,9 +351,9 @@ class FTXVodPlayConfigPlayerMsg {
 
   int? progressInterval;
 
-  int? maxBufferSize;
+  double? maxBufferSize;
 
-  int? maxPreloadSize;
+  double? maxPreloadSize;
 
   int? firstStartPlayBufferTime;
 
@@ -408,8 +408,8 @@ class FTXVodPlayConfigPlayerMsg {
       smoothSwitchBitrate: result[8] as bool?,
       cacheMp4ExtName: result[9] as String?,
       progressInterval: result[10] as int?,
-      maxBufferSize: result[11] as int?,
-      maxPreloadSize: result[12] as int?,
+      maxBufferSize: result[11] as double?,
+      maxPreloadSize: result[12] as double?,
       firstStartPlayBufferTime: result[13] as int?,
       nextStartPlayBufferTime: result[14] as int?,
       overlayKey: result[15] as String?,
@@ -768,7 +768,7 @@ class PreLoadMsg {
 
   String? playUrl;
 
-  int? preloadSizeMB;
+  double? preloadSizeMB;
 
   int? preferredResolution;
 
@@ -784,7 +784,7 @@ class PreLoadMsg {
     result as List<Object?>;
     return PreLoadMsg(
       playUrl: result[0] as String?,
-      preloadSizeMB: result[1] as int?,
+      preloadSizeMB: result[1] as double?,
       preferredResolution: result[2] as int?,
     );
   }
@@ -809,7 +809,7 @@ class PreLoadInfoMsg {
 
   String? playUrl;
 
-  int? preloadSizeMB;
+  double? preloadSizeMB;
 
   int? preferredResolution;
 
@@ -834,7 +834,7 @@ class PreLoadInfoMsg {
       fileId: result[1] as String?,
       pSign: result[2] as String?,
       playUrl: result[3] as String?,
-      preloadSizeMB: result[4] as int?,
+      preloadSizeMB: result[4] as double?,
       preferredResolution: result[5] as int?,
       tmpPreloadTaskId: result[6] as int?,
     );
